+++ CHANGELOG.md
@@ -28,6 +28,7 @@ For details about compatibility between different releases, see the **Commitment
 - Uplink and downlink counters display on end device activity in the Console.
 - Join settings handling in JS-only deployments in the Console.
 - Configuring Packet Broker listed option when Packet Broker Agent is configured with a Packet Broker tenant API key.
+- Contact info validation through the Account app.
 
 ### Security